requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1
django-bootstrap4==0.0.8
django-filter==2.1.0
django-toolbelt==0.0.1
djangorestframework==3.9.4
gunicorn==19.9.0
PyMySQL==0.8.0
router==0.1
sqlparse==0.3.0
static3==0.7.0
